compile_ready_no_segmentation

+++ FSI/FSI3/level1/source/turekDynamic_v2.f90
@@ -54,7 +54,7 @@ program main
    write (*, *) 'Total DOFs solved  :', nDofBC
 
    topLeftPt = nDofBC - (nElx*degEl + 1)*nDofPerNode + 1
-   probeDof = nDofBC - (0.5*nElx*degEl + 1)*nDofPerNode + [1, 2]
+   probeDof = nDofBC - int((0.5*nElx*degEl + 1)*nDofPerNode) + [1, 2]
 
    call compDynRes(tStart, tEnd, dt, probeDof)
 
@@ -338,12 +338,12 @@ contains
 
                !C=E/(1-v^2)*[1 v 0;v 1 0;0 0 (1-v)/2];
                C = E/((1 + v)*(1 - 2*v))*reshape([1 - v, v, 0.0d0, v, 1 - v, 0.0d0, 0.0d0, 0.0d0, (1 - 2*v)/2], &
-                                                 [3, 3], order=[2, 1])
+                  [3, 3], order=[2, 1])
 
                S01 = mulMatVec(C, E01)
 
                S_Mat = reshape([S01(1), S01(3), 0.0d0, 0.0d0, S01(3), S01(2), 0.0d0, 0.0d0, &
-                                0.0d0, 0.0d0, S01(1), S01(3), 0.0d0, 0.0d0, S01(3), S01(2)], [4, 4], order=[2, 1])
+                  0.0d0, 0.0d0, S01(1), S01(3), 0.0d0, 0.0d0, S01(3), S01(2)], [4, 4], order=[2, 1])
 
                BLt = transpose(BL)
                BLtCBL = Lz*mulMat(BLt, mulMat(C, BL)) !Integrand for KL
